+++ components/calendars/src/datepicker/datepicker-all.module.ts
@@ -2,10 +2,10 @@ import { NgModule, ValueProvider } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { DatePickerComponent } from './datepicker.component';
 import { DatePickerModule } from './datepicker.module';
+import {MaskedDateTime} from '@syncfusion/ej2-calendars'
 
 
-
-
+export const MaskedDateTimeService: ValueProvider = { provide: 'CalendarsMaskedDateTime', useValue: MaskedDateTime};
 
 /**
  * NgModule definition for the DatePicker component with providers.
@@ -16,7 +16,7 @@ import { DatePickerModule } from './datepicker.module';
         DatePickerModule
     ],
     providers:[
-        
+        MaskedDateTimeService
     ]
 })
 export class DatePickerAllModule { }