+++ components/calendars/src/index.ts
@@ -3,7 +3,7 @@ export { CalendarModule } from './calendar/calendar.module';
 export { CalendarAllModule, IslamicService } from './calendar/calendar-all.module';
 export { DatePickerComponent} from './datepicker/datepicker.component';
 export { DatePickerModule } from './datepicker/datepicker.module';
-export { DatePickerAllModule } from './datepicker/datepicker-all.module';
+export { DatePickerAllModule, MaskedDateTimeService } from './datepicker/datepicker-all.module';
 export { TimePickerComponent} from './timepicker/timepicker.component';
 export { TimePickerModule } from './timepicker/timepicker.module';
 export { TimePickerAllModule } from './timepicker/timepicker-all.module';