+++ components/inplaceeditor/src/inplace-editor/inplaceeditor.component.ts
@@ -6,7 +6,7 @@ import { Template } from '@syncfusion/ej2-angular-base';
 
 
 export const inputs: string[] = ['actionOnBlur','adaptor','cancelButton','cssClass','disabled','editableOn','emptyText','enableEditMode','enableHtmlParse','enableHtmlSanitizer','enablePersistence','enableRtl','locale','mode','model','name','popupSettings','primaryKey','saveButton','showButtons','submitOnEnter','template','textOption','type','url','validationRules','value'];
-export const outputs: string[] = ['focus', 'blur', 'actionBegin','actionFailure','actionSuccess','beforeSanitizeHtml','beginEdit','cancelClick','change','created','destroyed','submitClick','validating','valueChange'];
+export const outputs: string[] = ['focus', 'blur', 'actionBegin','actionFailure','actionSuccess','beforeSanitizeHtml','beginEdit','cancelClick','change','created','destroyed','endEdit','submitClick','validating','valueChange'];
 export const twoWays: string[] = ['value'];
 
 /**